% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{adjacency}
\alias{adjacency}
\alias{adjacency.default}
\alias{adjacency.rwclust}
\title{Generic helper for extracting adjacency matrix from rwclust object.}
\usage{
adjacency(x)

\method{adjacency}{default}(x)

\method{adjacency}{rwclust}(x)
}
\arguments{
\item{x}{rwclust object}
}
\value{
Matrix object containing the adjacency matrix of the after the final iteration
}
\description{
Generic helper for extracting adjacency matrix from rwclust object.
}
