% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metric_utils.R
\name{apply_similarity}
\alias{apply_similarity}
\title{Apply similarity function to rows of a matrix}
\usage{
apply_similarity(idx, mat, similarity, ...)
}
\arguments{
\item{idx}{vector of length two containing row indices}

\item{mat}{a matrix}

\item{similarity}{similarity function to apply}

\item{...}{additional parameters to be passed to the similarity function}
}
\value{
a scalar
}
\description{
Apply similarity function to rows of a matrix
}
