% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_functions.R
\name{compute_transition_matrix}
\alias{compute_transition_matrix}
\title{Compute transition matrix}
\usage{
compute_transition_matrix(x)
}
\arguments{
\item{x}{sparseMatrix or denseMatrix}
}
\value{
transition matrix
}
\description{
Compute transition matrix
}
