% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_functions.R
\name{create_weight_matrix}
\alias{create_weight_matrix}
\title{Construct sparse matrix from weighted edgelist}
\usage{
create_weight_matrix(edgelist, weights, ...)
}
\arguments{
\item{edgelist}{a dataframe with two columns}

\item{weights}{a vector of weights}

\item{...}{other parameters to be passed to Matrix::sparseMatrix()}
}
\value{
sparseMatrix
}
\description{
Takes the weights from compute_kernel and creates weighted adjacency matrix
}
