% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example1}
\alias{example1}
\title{Example Graph 1}
\format{
A data frame with three columns representing a weighted graph. Each
row represents an edge with a weight:
\describe{
\item{from}{An integer vertex id}
\item{to}{An integer vertex id}
\item{weight}{A double representing the edge weight}
}
}
\usage{
example1
}
\description{
First demonstration test graph used in the original.
}
\examples{
data(example1, package="Rwclust")

}
\keyword{datasets}
