% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example2}
\alias{example2}
\title{Example Graph 2}
\format{
A data frame with three columns representing a weighted graph. Each
row represents an edge with a weight.
\describe{
\item{from}{An integer vertex id}
\item{to}{An integer vertex id}
\item{weight}{A double representing the edge weight}
}
}
\source{
\url{https://link.springer.com/chapter/10.1007/3-540-45294-X_3}
}
\usage{
example2
}
\description{
Second demonstration test graph used in the original paper by Harel and Koren.
}
\examples{
\dontrun{
 data(example2, package="Rwclust")
}

}
\references{
Harel, David, and Yehuda Koren. "On clustering using random walks."
International Conference on Foundations of Software Technology
and Theoretical Computer Science. Springer, Berlin, Heidelberg, 2001.
}
\keyword{datasets}
