% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{new_rwclust}
\alias{new_rwclust}
\title{rwclust class constructor}
\usage{
new_rwclust(weights, adj)
}
\arguments{
\item{weights}{vector of double containing the edge weights}

\item{adj}{Matrix produced at the final iteration of the algorithm}
}
\description{
Returns a object of class "rwclust" for use with generic summary and plotting functions.
}
