% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plot.rwclust}
\alias{plot.rwclust}
\title{Generic plotting for rwclust object}
\usage{
\method{plot}{rwclust}(x, cutoff = NULL, ...)
}
\arguments{
\item{x}{rwclust object}

\item{cutoff}{optional numeric, will plot the cutoff value as a vertical line}

\item{...}{additional graphical parameters passed to the \code{hist} function}
}
\description{
Generic function for plotting the distribution of weights. Calls \code{hist} under the hood.
}
