% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_functions.R
\name{run_main_loop}
\alias{run_main_loop}
\title{Execute main algorithm loop}
\usage{
run_main_loop(M, edgelist, similarity, k, iter)
}
\arguments{
\item{M}{transition matrix}

\item{edgelist}{dataframe edgelist}

\item{similarity}{a similarity function}

\item{k}{integer, length of longest walk}

\item{iter}{number of iterations}
}
\value{
list
}
\description{
Execute main algorithm loop
}
