% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rwclust.R
\name{rwclust_}
\alias{rwclust_}
\title{Executes the full algorithm}
\usage{
rwclust_(adj, edgelist, similarity = "hk", iter, k)
}
\arguments{
\item{adj}{a sparseMatrix representing the adjacency matrix}

\item{edgelist}{dataframe edgelist}

\item{similarity}{string, the name of the similarity metric used to
update weights}

\item{iter}{integer, number of iterations}

\item{k}{integer, maximum length of random walk}
}
\value{
list
\describe{
\item{weights}{A vector of the updated edge weights}
\item{adj}{Updated adjacency matrix containing updated weights}
}
}
\description{
\code{rwclust} servers as the user-facing wrapper function for this function
}
