% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_functions.R
\name{update_weights}
\alias{update_weights}
\title{Update edge weights}
\usage{
update_weights(M, el, similarity, k)
}
\arguments{
\item{M}{matrix}

\item{el}{dataframe representing weighted edgelist}

\item{similarity}{a similarity function}

\item{k}{integer, length of longest walk}
}
\description{
Update edge weights
}
