% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example2}
\alias{example2}
\title{Example Graph 2}
\format{
A data frame with three columns representing a weighted graph. Each
row represents an edge with a weight.
\describe{
\item{from}{An integer vertex id}
\item{to}{An integer vertex id}
\item{weight}{A double representing the edge weight}
}
}
\usage{
example2
}
\description{
Second demonstration test graph used in the original paper.
}
\examples{
data(example2, package="Rwclust")

}
\keyword{datasets}
