% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{new_rwclust}
\alias{new_rwclust}
\title{rwclust class constructor}
\usage{
new_rwclust(x)
}
\arguments{
\item{x}{output of \code{run_main_loop} function}
}
\description{
Returns a object of class "rwclust" for use with generic summary and plotting functions.
}
\seealso{
\code{\link[=run_main_loop]{run_main_loop()}}
}
